add active failures to comfort mode
