+++ src/instruments/src/EFB/Failures/Pages/Comfort/index.tsx
@@ -6,6 +6,9 @@ import { Failure } from '@flybywiresim/failures';
 import { pathify } from '../../../Utils/routing';
 import { AtaChapterPage } from './AtaChapterPage';
 import { useFailuresOrchestrator } from '../../../failures-orchestrator-provider';
+import { FailureGroup } from '../FailureGroup';
+import { ScrollableContainer } from '../../../UtilComponents/ScrollableContainer';
+import { useAppSelector } from '../../../Store/store';
 
 interface ATAChapterCardProps {
     ataNumber: AtaChapterNumber,
@@ -13,43 +16,27 @@ interface ATAChapterCardProps {
     description: string,
 }
 
-const ATAChapterCard = ({ ataNumber, description, title }: ATAChapterCardProps) => {
-    const { activeFailures, allFailures } = useFailuresOrchestrator();
-
-    const hasActiveFailure = allFailures
-        .filter((it) => it.ata === ataNumber)
-        .some((it) => activeFailures.has(it.identifier));
-
-    return (
-        <Link
-            to={`/failures/comfort/${pathify(ataNumber.toString())}`}
-            className="flex flex-row p-2 space-x-4 rounded-md border-2 border-transparent transition duration-100 hover:border-theme-highlight"
+const ATAChapterCard = ({ ataNumber, description, title }: ATAChapterCardProps) => (
+    <Link
+        to={`/failures/comfort/${pathify(ataNumber.toString())}`}
+        className="flex flex-row p-2 space-x-4 rounded-md border-2 border-transparent hover:border-theme-highlight transition duration-100"
+    >
+        <div
+            className="flex justify-center items-center w-1/5 font-title text-5xl font-bold bg-theme-accent rounded-md"
         >
-            <div
-                className="flex justify-center items-center w-1/5 text-5xl font-bold rounded-md font-title bg-theme-accent"
-            >
-                {`ATA ${ataNumber}`}
+            {`ATA ${ataNumber}`}
+        </div>
 
-                <div className="inline-block relative -right-7 bottom-16 w-0 h-0 fill-current text-utility-red">
-                    {hasActiveFailure && (
-                        <svg style={{ width: '30px', height: '30px' }} viewBox="0 0 20 20">
-                            <circle cx={10} cy={10} r={5} />
-                        </svg>
-                    )}
-                </div>
-            </div>
-
-            <div className="space-y-2 w-3/4">
-                <h1 className="font-bold">
-                    {title}
-                </h1>
-                <p>
-                    {description}
-                </p>
-            </div>
-        </Link>
-    );
-};
+        <div className="space-y-2 w-3/4">
+            <h1 className="font-bold">
+                {title}
+            </h1>
+            <p>
+                {description}
+            </p>
+        </div>
+    </Link>
+);
 
 interface ComfortUIProps {
     filteredChapters: AtaChapterNumber[];
@@ -57,22 +44,33 @@ interface ComfortUIProps {
     failures: Failure[];
 }
 
-export const ComfortUI = ({ filteredChapters, allChapters, failures }: ComfortUIProps) => (
-    <>
-        <Route exact path="/failures/comfort">
-            {filteredChapters.map((chapter) => (
-                <ATAChapterCard
-                    ataNumber={chapter}
-                    title={AtaChaptersTitle[chapter]}
-                    description={AtaChaptersDescription[chapter]}
-                />
-            ))}
-        </Route>
+export const ComfortUI = ({ filteredChapters, allChapters, failures }: ComfortUIProps) => {
+    const { allFailures, activeFailures } = useFailuresOrchestrator();
+    const { searchQuery } = useAppSelector((state) => state.failuresPage);
 
-        {allChapters.map((chapter) => (
-            <Route path={`/failures/comfort/${chapter.toString()}`}>
-                <AtaChapterPage chapter={chapter} failures={failures} />
+    return (
+        <>
+            <Route exact path="/failures/comfort">
+                <div className="space-y-8">
+                    {searchQuery.length === 0 && activeFailures.size !== 0 && (
+                        <FailureGroup className="ml-2.5" title="Active Failures" failures={allFailures.filter((failure) => activeFailures.has(failure.identifier))} />
+                    )}
+                </div>
+                <ScrollableContainer height={30}>
+                    {filteredChapters.map((chapter) => (
+                        <ATAChapterCard
+                            ataNumber={chapter}
+                            title={AtaChaptersTitle[chapter]}
+                            description={AtaChaptersDescription[chapter]}
+                        />
+                    ))}
+                </ScrollableContainer>
             </Route>
-        ))}
-    </>
-);
+            {allChapters.map((chapter) => (
+                <Route path={`/failures/comfort/${chapter.toString()}`}>
+                    <AtaChapterPage chapter={chapter} failures={failures} />
+                </Route>
+            ))}
+        </>
+    );
+};
