+++ src/instruments/src/EFB/Failures/Pages/Compact.tsx
@@ -1,55 +1,10 @@
 import { AtaChapterNumber, AtaChaptersTitle } from '@shared/ata';
 import React from 'react';
 import { Failure } from '@flybywiresim/failures';
-import { FailureButton } from '../FailureButton';
 import { useFailuresOrchestrator } from '../../failures-orchestrator-provider';
 import { useAppSelector } from '../../Store/store';
 import { ScrollableContainer } from '../../UtilComponents/ScrollableContainer';
-
-interface FailureGroupProps {
-    title: string;
-    failures: Failure[];
-}
-
-const FailureGroup = ({ title, failures }: FailureGroupProps) => {
-    const { activeFailures, changingFailures, activate, deactivate } = useFailuresOrchestrator();
-    const { searchQuery } = useAppSelector((state) => state.failuresPage);
-
-    const getHighlightedTerm = (failureName: string) => {
-        const searchQueryIdx = failureName.toUpperCase().indexOf(searchQuery);
-
-        if (searchQuery === '' || searchQueryIdx === -1) return undefined;
-
-        return failureName.substring(searchQueryIdx, searchQueryIdx + searchQuery.length);
-    };
-
-    const handleFailureButtonClick = (failureIdentifier: number) => {
-        if (!activeFailures.has(failureIdentifier)) {
-            activate(failureIdentifier);
-        } else {
-            deactivate(failureIdentifier);
-        }
-    };
-
-    return (
-        <div className="space-y-2">
-            <h2>{title}</h2>
-
-            <div className="grid grid-cols-4 auto-rows-auto">
-                {failures.map((failure, index) => (
-                    <FailureButton
-                        name={failure.name}
-                        isActive={activeFailures.has(failure.identifier)}
-                        isChanging={changingFailures.has(failure.identifier)}
-                        highlightedTerm={getHighlightedTerm(failure.name)}
-                        onClick={() => handleFailureButtonClick(failure.identifier)}
-                        className={`${index && index % 4 !== 0 && 'ml-4'} ${index >= 4 && 'mt-4'} h-36`}
-                    />
-                ))}
-            </div>
-        </div>
-    );
-};
+import { FailureGroup } from './FailureGroup';
 
 interface CompactUIProps {
     chapters: AtaChapterNumber[];
